requirements.txt
fastapi==0.104.0
uvicorn[standard]
SQLAlchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
passlib==1.7.4
python-multipart==0.0.5
bcrypt==4.0.1
cryptography
pathlib==1.0.1
fastapi-mail==1.4.1
requests
boto3==1.34.84
redis==5.0.6
celery==5.4.0
aio-pika==9.4.3
celery[redis]
celery[beat]
flower==2.0.1
pandas==2.2.3
Pillow==11.0.0
Pygments==2.18.0
Requests==2.32.3
scikit_learn==1.4.2
sentence_transformers==3.3.1
Sphinx==8.1.3
sphinx_rtd_theme==3.0.2
torch==2.5.1
transformers==4.46.3